add ShopService.java and refactor controller code to service class
