+++ src/main/java/se/iths/worldfirstwebshop/webshop/access/Shop.java
@@ -7,7 +7,6 @@ import se.iths.worldfirstwebshop.webshop.product.Product;
 import se.iths.worldfirstwebshop.webshop.storage.Cart;
 import se.iths.worldfirstwebshop.webshop.storage.Inventory;
 
-@Component
 public class Shop {
 
     private final Cart cart;
